refactor: split gamelibraryview into hero card, search bar, sort sheet, sorting helpers
